+++ src/components/Auth/auth.css
@@ -5,7 +5,7 @@
   border-radius: 20px;
   padding: 10px;
   max-width: 500px;
-  margin: 150px auto;
+  margin: 150px auto 0;
   font-size: 18px;
 }
 .auth-error {
