+++ src/index.tsx
@@ -7,26 +7,25 @@ import { App } from './App';
 import {
     Games, Manual, Statistic, Team,
 } from './routes';
-import { AuthForm, Footer, ResponsiveAppBar } from './components';
-import { Provider } from './core/context';
+import { Footer, ResponsiveAppBar, Authentication } from './components';
+import { LogIn } from './components/Auth/LogIn';
 
 const rootElement = document.getElementById('root');
 
 render(
     <BrowserRouter>
         <div className="App">
-            <Provider value={localStorage.getItem('token')}>
-                <ResponsiveAppBar />
-                <Routes>
-                    <Route path="/" element={<App />} />
-                    <Route path="auth" element={<AuthForm />} />
-                    <Route path="games" element={<Games />} />
-                    <Route path="manual" element={<Manual />} />
-                    <Route path="statistic" element={<Statistic />} />
-                    <Route path="team" element={<Team />} />
-                </Routes>
-                <Footer />
-            </Provider>
+            <ResponsiveAppBar />
+            <Routes>
+                <Route path="/" element={<App />} />
+                <Route path="auth" element={<Authentication />} />
+                <Route path="login" element={<LogIn />} />
+                <Route path="games" element={<Games />} />
+                <Route path="manual" element={<Manual />} />
+                <Route path="statistic" element={<Statistic />} />
+                <Route path="team" element={<Team />} />
+            </Routes>
+            <Footer />
         </div>
     </BrowserRouter>,
     rootElement,
